FIXED: This should fix the linux abs(float) ambigious call error.

+++ src/arcemu-shared/LocationVector.h
@@ -165,9 +165,9 @@ public:
 
 	bool operator == (const LocationVector & eq)
 	{
-		float dx = abs(eq.x - x);
-		float dy = abs(eq.y - y);
-		float dz = abs(eq.z - z);
+		float dx = abs((float)eq.x - x);
+		float dy = abs((float)eq.y - y);
+		float dz = abs((float)eq.z - z);
 		float e = 0.0001f;
 		bool result = false;
 		if(dx <= e && dy <= e && dz <= e)
@@ -176,9 +176,9 @@ public:
 	}
 	bool operator != (const LocationVector & loc)
 	{
-		float dx = abs(loc.x - x);
-		float dy = abs(loc.y - y);
-		float dz = abs(loc.z - z);
+		float dx = abs((float)loc.x - x);
+		float dy = abs((float)loc.y - y);
+		float dz = abs((float)loc.z - z);
 		float e = 0.0001f;
 		bool result = false;
 		//If the difference between either coord and their counter part is significant.
